(modelLoader) init

+++ game/game.go
@@ -12,5 +12,5 @@ type Game interface {
 	IsGameFinished() (bool, PlayerColor)
 	DrawBoard()
 	SerializeBoard() string
-	OneHotBoard() [][][]int8
+	OneHotBoard() [][][]float32
 }
